+++ lib/instream/pool/worker.ex
@@ -11,19 +11,19 @@ defmodule Instream.Pool.Worker do
 
   @behaviour :poolboy_worker
 
-  def start_link(conn) do
-    GenServer.start_link(__MODULE__, conn)
+  def start_link(default) do
+    GenServer.start_link(__MODULE__, default)
   end
 
-  def init(conn) do
-    case conn[:writer] do
-      Instream.Writer.UDP -> { :ok, connect_udp(conn) }
-      _                   -> { :ok, conn }
+  def init(%{ module: conn } = default) do
+    case conn.config([ :writer ]) do
+      Instream.Writer.UDP -> { :ok, connect_udp(default) }
+      _                   -> { :ok, default }
     end
   end
 
-  def terminate(_reason, conn) do
-    case conn[:udp_socket] do
+  def terminate(_reason, state) do
+    case Map.get(state, :udp_socket) do
       nil    -> :ok
       socket -> :gen_udp.close(socket)
     end
@@ -32,32 +32,31 @@ defmodule Instream.Pool.Worker do
 
   # GenServer callbacks
 
-  def handle_call({ :execute, query, opts }, _from, conn) do
-    { :reply, execute(query, opts, conn), conn }
+  def handle_call({ :execute, query, opts }, _from, state) do
+    { :reply, execute(query, opts, state), state }
   end
 
-  def handle_cast({ :execute, query, opts }, conn) do
-    execute(query, opts, conn)
+  def handle_cast({ :execute, query, opts }, state) do
+    execute(query, opts, state)
 
-    { :noreply, conn }
+    { :noreply, state }
   end
 
 
   # Utility methods
 
-  defp connect_udp(conn) do
+  defp connect_udp(state) do
     { :ok, socket } = :gen_udp.open(0, [ :binary, { :active, false }])
 
-    conn
-    |> Keyword.put_new(:udp_socket, socket)
+    Map.put(state, :udp_socket, socket)
   end
 
-  defp execute(%Query{ type: type } = query, opts, conn) do
+  defp execute(%Query{ type: type } = query, opts, state) do
     case type do
-      :ping   -> QueryRunner.ping(query, opts, conn)
-      :read   -> QueryRunner.read(query, opts, conn)
-      :status -> QueryRunner.status(query, opts, conn)
-      :write  -> QueryRunner.write(query, opts, conn)
+      :ping   -> QueryRunner.ping(query, opts, state)
+      :read   -> QueryRunner.read(query, opts, state)
+      :status -> QueryRunner.status(query, opts, state)
+      :write  -> QueryRunner.write(query, opts, state)
     end
   end
 end